refactor(lexer): change SToken m_id field datatype to unsigned long

+++ src/lexer/lexer.cpp
@@ -68,7 +68,7 @@ namespace cherubim::lexer
 
     vector<SToken> lex(string input)
     {
-        unsigned idCounter;
+        unsigned long idCounter;
         vector<SToken> tokens;
         vector<string> splitInput;
